+++ components/Header/MobileMenuPopup.jsx
@@ -8,8 +8,8 @@ import {
 	UserIcon,
 	WalletIcon,
 } from '@heroicons/react/24/solid'
-import { useAuth } from '../../hooks/auth'
 import Link from 'next/link'
+import { useAuth } from 'hooks/auth'
 
 export default () => {
 	const { user, logout } = useAuth()
@@ -35,9 +35,8 @@ export default () => {
 									{({ active }) => (
 										<Link
 											href="/user/dashboard"
-											className={`${
-												active && 'bg-slate-100'
-											} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
+											className={`${active && 'bg-slate-100'
+												} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
 											<UserIcon className="mr-2 w-4 text-gray-500" />
 											Account
 										</Link>
@@ -48,9 +47,8 @@ export default () => {
 								<Menu.Item>
 									{({ active }) => (
 										<button
-											className={`${
-												active && 'bg-slate-100'
-											} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
+											className={`${active && 'bg-slate-100'
+												} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
 											<WalletIcon className="mr-2 w-4 text-gray-500" />
 											Add Funds
 										</button>
@@ -61,9 +59,8 @@ export default () => {
 								<Menu.Item>
 									{({ active }) => (
 										<button
-											className={`${
-												active && 'bg-slate-100'
-											} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
+											className={`${active && 'bg-slate-100'
+												} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
 											<QueueListIcon className="mr-2 w-4 text-gray-500" />
 											Orders
 										</button>
@@ -72,9 +69,8 @@ export default () => {
 								<Menu.Item>
 									{({ active }) => (
 										<button
-											className={`${
-												active && 'bg-slate-100'
-											} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
+											className={`${active && 'bg-slate-100'
+												} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
 											<BellIcon className="mr-2 w-4 text-gray-500" />
 											Notification
 										</button>
@@ -86,9 +82,8 @@ export default () => {
 									{({ active }) => (
 										<button
 											onClick={() => logout()}
-											className={`${
-												active && 'bg-slate-100'
-											} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
+											className={`${active && 'bg-slate-100'
+												} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
 											<PowerIcon className="mr-2 w-4 text-gray-500" />
 											Logout
 										</button>
@@ -102,9 +97,8 @@ export default () => {
 								{({ active }) => (
 									<Link
 										href="/register"
-										className={`${
-											active && 'bg-slate-100'
-										} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
+										className={`${active && 'bg-slate-100'
+											} group flex w-full items-center rounded-md px-2 py-2 text-sm text-gray-500`}>
 										<UserIcon className="mr-2 w-4 text-gray-500" />
 										Sign up
 									</Link>
