+++ components/Nav.jsx
@@ -4,6 +4,7 @@ import Link from 'next/link'
 import IconEnvelopeCircle from './Icons/IconEnvelopeCircle'
 import IconPhoneCircle from './Icons/IconPhoneCircle'
 
+
 export default () => {
 	const NavLink = ({ href, cta, className, children, ...props }) => {
 		return (
