+++ components/PaymentMethods.jsx
@@ -1,7 +1,7 @@
 import { Fragment, useEffect, useState } from 'react'
 import { Listbox, Transition } from '@headlessui/react'
 import { CheckIcon, ChevronUpDownIcon } from '@heroicons/react/20/solid'
-import axios from '../lib/axios'
+import axios from 'lib/axios'
 
 export default ({ setShowCardForm }) => {
 	const [methods, setMethods] = useState([])
@@ -45,25 +45,23 @@ export default ({ setShowCardForm }) => {
 					leaveFrom="opacity-100"
 					leaveTo="opacity-0">
 					<Listbox.Options className="absolute mt-1 max-h-60 w-full overflow-auto rounded-md bg-white py-1 text-base shadow-lg ring-1 ring-black ring-opacity-5 focus:outline-none sm:text-sm">
-						{methods.map((method, personIdx) => (
+						{methods.length >= 0 && methods?.map((method, personIdx) => (
 							<Listbox.Option
 								key={personIdx}
 								className={({ active }) =>
-									`relative cursor-default select-none py-2 pl-10 pr-4 ${
-										active
-											? 'bg-amber-100 text-amber-900'
-											: 'text-gray-900'
+									`relative cursor-default select-none py-2 pl-10 pr-4 ${active
+										? 'bg-amber-100 text-amber-900'
+										: 'text-gray-900'
 									}`
 								}
 								value={method.cardHolder}>
 								{({ selected }) => (
 									<>
 										<span
-											className={`block truncate ${
-												selected
-													? 'font-medium'
-													: 'font-normal'
-											}`}>
+											className={`block truncate ${selected
+												? 'font-medium'
+												: 'font-normal'
+												}`}>
 											{method.cardHolder}
 										</span>
 										{selected ? (
@@ -80,21 +78,19 @@ export default ({ setShowCardForm }) => {
 						))}
 						<Listbox.Option
 							className={({ active }) =>
-								`relative cursor-default select-none py-2 pl-10 pr-4 ${
-									active
-										? 'bg-amber-100 text-amber-900'
-										: 'text-gray-900'
+								`relative cursor-default select-none py-2 pl-10 pr-4 ${active
+									? 'bg-amber-100 text-amber-900'
+									: 'text-gray-900'
 								}`
 							}
 							value="new">
 							{({ selected }) => (
 								<>
 									<span
-										className={`block truncate ${
-											selected
-												? 'font-medium'
-												: 'font-normal'
-										}`}>
+										className={`block truncate ${selected
+											? 'font-medium'
+											: 'font-normal'
+											}`}>
 										+ Add New Payment Method
 									</span>
 									{selected ? (
